Specify that name constraints are only for a leaf.
And when you need these, you can ensure you make the issueing
intermediate be the place they are.

And you do not need to make this apply up a chain.. because that
is terrible, and since this is an otherName, 5280 lets us make
the rules.

I believe this both makes something "sufficiently name constrait-y"
to be useful for those of us (including myself in my past life)
who would have made use of this, while making this not horrible
to implement. Yes, this means that anyone wishing to use this
must put them into the leaf issuer, and can't encode their org
chart in a chain. IMO, this is fine.